+++ CHANGELOG.md
@@ -6,6 +6,9 @@ To see compatibility information, `xsvutils help compatibility`.
 
 ## v5 -> v6 (Unreleased)
 
+以下のコマンドが追加された。
+- inslinenum
+
 以下のオプションが追加された。
 - grep コマンドの -v
 
@@ -67,6 +70,7 @@ insconst のパラメータの順序が逆になってしまった。 (see `xsvu
 - add `-v` option of `grep` subcommand
 - fix a bug of handling division by zero on `facetcount` subcommand
 - improve handling empty input
+- add `inslinenum` subcommand
 
 ### version 0.5 (-v5) (2018/07/16)
 - add support input of xz file